Implemented multi-step puzzle
Implemented section of game where:
Player finds a locked door.
They must look into a nearby room to find the code for a safe.
In the safe, there is an AudioRecording.
They bring the recording to the Grammophone.
They play the recording to learn the code to the locked door.