The Push of the Day
Did some much needed tweaks to the Game Scene and made the quality of life easier for the game. The Cassettes have been placed in general locations that made sense to me. We can go over it as a team next we meet. Fixed a bunch of key collisions and placed them where they needed to be placed. Lighting is sorted although going to hold on baking light into the Game Scene until we are more fully ready for it.